move fixtures.py to test_warehouses.py
